Start separation of rendering logic
Took 19 minutes
