+++ .idea/dream.iml
@@ -7,6 +7,7 @@
       <sourceFolder url="file://$MODULE_DIR$/dream-resource/src" isTestSource="false" />
       <sourceFolder url="file://$MODULE_DIR$/src" isTestSource="false" />
       <sourceFolder url="file://$MODULE_DIR$/dream-application/src" isTestSource="false" />
+      <sourceFolder url="file://$MODULE_DIR$/dream-renderer/src" isTestSource="false" />
       <excludeFolder url="file://$MODULE_DIR$/target" />
       <excludeFolder url="file://$MODULE_DIR$/dream-resource/target" />
       <excludeFolder url="file://$MODULE_DIR$/dream-editor/target" />
