+++ dream-renderer/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "dream"
+name = "dream-renderer"
 version = "0.1.0"
 edition = "2021"
 
@@ -7,8 +7,7 @@ edition = "2021"
 crate-type = ["cdylib", "rlib"]
 
 [dependencies]
-dream-editor = { path = "./dream-editor", version = "0.1.0" }
-dream-renderer = { path = "./dream-renderer", version = "0.1.0" }
+dream-editor = { path = "../dream-editor", version = "0.1.0" }
 instant = { version = "0.1", features = [ "stdweb" ] }
 cfg-if = "1"
 winit = "0.28"
@@ -46,9 +45,3 @@ web-sys = { version = "0.3", features = [
     "Window",
     "Element",
 ]}
-
-[workspace]
-members = [
-    "dream-editor",
-    "dream-renderer"
-]
