+++ dream-renderer/src/happy-tree.png
