+++ dream-renderer/src/icons/EditIconDark.png
