+++ dream-renderer/src/icons/EditIconLight.png
