+++ dream-renderer/src/icons/ExpandIcon.png
