+++ dream-renderer/src/icons/PlayIconDark.png
