+++ dream-renderer/src/icons/PlayIconLight.png
