+++ dream-renderer/src/icons/SelectIconBlack.png
