+++ dream-renderer/src/icons/SelectIconWhite.png
