+++ dream-renderer/src/icons/StopIcon.png
